refactor: split admin.py
