--- backend/app/api/endpoints/admin.py
@@ -1,2863 +1,0 @@
-# SPDX-FileCopyrightText: 2025 Weibo, Inc.
-#
-# SPDX-License-Identifier: Apache-2.0
-
-import asyncio
-import threading
-from typing import Any, Dict, List, Optional
-
-from fastapi import APIRouter, Depends, HTTPException, Path, Query, status
-from sqlalchemy import func
-from sqlalchemy.orm import Session
-
-from app.api.dependencies import get_db
-from app.api.endpoints.kind.common import (
-    format_resource_list,
-    format_single_resource,
-    prepare_batch_resources,
-    validate_and_prepare_resource,
-    validate_resource_type,
-    validate_user_exists,
-)
-from app.api.endpoints.kind.kinds import KIND_SCHEMA_MAP
-from app.core.security import create_access_token, get_admin_user, get_password_hash
-from app.models.kind import Kind
-from app.models.system_config import SystemConfig
-from app.models.user import User
-from app.schemas.admin import (
-    AdminUserCreate,
-    AdminUserListResponse,
-    AdminUserResponse,
-    AdminUserUpdate,
-    ChatSloganItem,
-    ChatSloganTipsResponse,
-    ChatSloganTipsUpdate,
-    ChatTipItem,
-    PasswordReset,
-    PublicBotCreate,
-    PublicBotListResponse,
-    PublicBotResponse,
-    PublicBotUpdate,
-    PublicGhostCreate,
-    PublicGhostListResponse,
-    PublicGhostResponse,
-    PublicGhostUpdate,
-    PublicModelCreate,
-    PublicModelListResponse,
-    PublicModelResponse,
-    PublicModelUpdate,
-    PublicRetrieverListResponse,
-    PublicRetrieverResponse,
-    PublicShellCreate,
-    PublicShellListResponse,
-    PublicShellResponse,
-    PublicShellUpdate,
-    PublicTeamCreate,
-    PublicTeamListResponse,
-    PublicTeamResponse,
-    PublicTeamUpdate,
-    QuickAccessResponse,
-    QuickAccessTeam,
-    RoleUpdate,
-    SystemConfigResponse,
-    SystemConfigUpdate,
-    SystemStats,
-)
-from app.schemas.kind import BatchResponse, Retriever
-from app.schemas.task import TaskCreate, TaskInDB
-from app.schemas.user import Token, UserInDB, UserInfo
-from app.services.adapters.public_retriever import public_retriever_service
-from app.services.adapters.task_kinds import task_kinds_service
-from app.services.k_batch import apply_default_resources_async, batch_service
-from app.services.kind import kind_service
-from app.services.user import user_service
-
-router = APIRouter()
-
-
-# ==================== User Management Endpoints ====================
-
-
-@router.get("/users", response_model=AdminUserListResponse)
-async def list_all_users(
-    page: int = Query(1, ge=1),
-    limit: int = Query(20, ge=1, le=100),
-    include_inactive: bool = Query(False),
-    search: Optional[str] = Query(None, description="Search by username or email"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get list of all users with pagination and search
-    """
-    query = db.query(User)
-    if not include_inactive:
-        query = query.filter(User.is_active == True)
-
-    # Apply search filter
-    if search:
-        search_pattern = f"%{search}%"
-        query = query.filter(
-            (User.user_name.ilike(search_pattern)) | (User.email.ilike(search_pattern))
-        )
-
-    total = query.count()
-    users = query.offset((page - 1) * limit).limit(limit).all()
-
-    return AdminUserListResponse(
-        total=total,
-        items=[
-            AdminUserResponse(
-                id=user.id,
-                user_name=user.user_name,
-                email=user.email,
-                role=user.role,
-                auth_source=user.auth_source,
-                is_active=user.is_active,
-                created_at=user.created_at,
-                updated_at=user.updated_at,
-            )
-            for user in users
-        ],
-    )
-
-
-@router.get("/users/{user_id}", response_model=AdminUserResponse)
-async def get_user_by_id_endpoint(
-    user_id: int = Path(..., description="User ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get detailed information for specified user ID
-    """
-    user = user_service.get_user_by_id(db, user_id)
-    return AdminUserResponse(
-        id=user.id,
-        user_name=user.user_name,
-        email=user.email,
-        role=user.role,
-        auth_source=user.auth_source,
-        is_active=user.is_active,
-        created_at=user.created_at,
-        updated_at=user.updated_at,
-    )
-
-
-@router.post(
-    "/users", response_model=AdminUserResponse, status_code=status.HTTP_201_CREATED
-)
-async def create_user(
-    user_data: AdminUserCreate,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Create a new user (admin only)
-    """
-    # Check if username already exists
-    existing_user = db.query(User).filter(User.user_name == user_data.user_name).first()
-    if existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"User with username '{user_data.user_name}' already exists",
-        )
-
-    # Validate password for password auth source
-    if user_data.auth_source == "password" and not user_data.password:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Password is required for password authentication",
-        )
-
-    # Create user
-    password_hash = (
-        get_password_hash(user_data.password)
-        if user_data.password
-        else get_password_hash("oidc_placeholder")
-    )
-    new_user = User(
-        user_name=user_data.user_name,
-        email=user_data.email,
-        password_hash=password_hash,
-        role=user_data.role,
-        auth_source=user_data.auth_source,
-        is_active=True,
-        git_info=[],
-        preferences="{}",
-    )
-    db.add(new_user)
-    db.commit()
-    db.refresh(new_user)
-
-    # Apply default resources for the new user in a background thread
-    def run_async_task():
-        asyncio.run(apply_default_resources_async(new_user.id))
-
-    thread = threading.Thread(target=run_async_task, daemon=True)
-    thread.start()
-
-    return AdminUserResponse(
-        id=new_user.id,
-        user_name=new_user.user_name,
-        email=new_user.email,
-        role=new_user.role,
-        auth_source=new_user.auth_source,
-        is_active=new_user.is_active,
-        created_at=new_user.created_at,
-        updated_at=new_user.updated_at,
-    )
-
-
-@router.put("/users/{user_id}", response_model=AdminUserResponse)
-async def update_user(
-    user_data: AdminUserUpdate,
-    user_id: int = Path(..., description="User ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Update user information (admin only)
-    """
-    # Query user directly to avoid decrypt_user_git_info modifying the object
-    # which can cause SQLAlchemy session state issues
-    user = db.query(User).filter(User.id == user_id).first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"User with id {user_id} not found",
-        )
-
-    # Prevent admin from deactivating themselves
-    if user.id == current_user.id and user_data.is_active is False:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Cannot deactivate your own account",
-        )
-
-    # Prevent admin from demoting themselves
-    if user.id == current_user.id and user_data.role == "user":
-        # Check if there are other admins
-        admin_count = (
-            db.query(User).filter(User.role == "admin", User.is_active == True).count()
-        )
-        if admin_count <= 1:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="Cannot demote yourself when you are the only admin",
-            )
-
-    # Check username uniqueness if being changed
-    if user_data.user_name and user_data.user_name != user.user_name:
-        existing_user = (
-            db.query(User).filter(User.user_name == user_data.user_name).first()
-        )
-        if existing_user:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail=f"User with username '{user_data.user_name}' already exists",
-            )
-
-    # Update fields
-    if user_data.user_name is not None:
-        user.user_name = user_data.user_name
-    if user_data.email is not None:
-        user.email = user_data.email
-    if user_data.role is not None:
-        user.role = user_data.role
-    if user_data.is_active is not None:
-        user.is_active = user_data.is_active
-
-    db.commit()
-    db.refresh(user)
-
-    return AdminUserResponse(
-        id=user.id,
-        user_name=user.user_name,
-        email=user.email,
-        role=user.role,
-        auth_source=user.auth_source,
-        is_active=user.is_active,
-        created_at=user.created_at,
-        updated_at=user.updated_at,
-    )
-
-
-@router.delete("/users/{user_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_user(
-    user_id: int = Path(..., description="User ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Delete a user (hard delete - permanently removes the user from database)
-    """
-    # Query user directly to avoid decrypt_user_git_info modifying the object
-    user = db.query(User).filter(User.id == user_id).first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"User with id {user_id} not found",
-        )
-
-    # Prevent admin from deleting themselves
-    if user.id == current_user.id:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Cannot delete your own account",
-        )
-
-    # Hard delete - permanently remove the user
-    db.delete(user)
-    db.commit()
-
-    return None
-
-
-@router.post("/users/{user_id}/reset-password", response_model=AdminUserResponse)
-async def reset_user_password(
-    password_data: PasswordReset,
-    user_id: int = Path(..., description="User ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Reset user password (admin only)
-    """
-    # Query user directly to avoid decrypt_user_git_info modifying the object
-    user = db.query(User).filter(User.id == user_id).first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"User with id {user_id} not found",
-        )
-
-    # Only allow password reset for non-OIDC users
-    # - "password": user registered with password
-    # - "unknown": legacy users before auth_source was added
-    # - "api:*": users auto-created via API service key
-    can_reset = user.auth_source in [
-        "password",
-        "unknown",
-    ] or user.auth_source.startswith("api:")
-    if not can_reset:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Cannot reset password for OIDC-authenticated users",
-        )
-
-    user.password_hash = get_password_hash(password_data.new_password)
-    if user.auth_source == "unknown":
-        user.auth_source = "password"
-    db.commit()
-    db.refresh(user)
-
-    return AdminUserResponse(
-        id=user.id,
-        user_name=user.user_name,
-        email=user.email,
-        role=user.role,
-        auth_source=user.auth_source,
-        is_active=user.is_active,
-        created_at=user.created_at,
-        updated_at=user.updated_at,
-    )
-
-
-@router.post("/users/{user_id}/toggle-status", response_model=AdminUserResponse)
-async def toggle_user_status(
-    user_id: int = Path(..., description="User ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Toggle user active status (enable/disable)
-    """
-    # Query user directly to avoid decrypt_user_git_info modifying the object
-    user = db.query(User).filter(User.id == user_id).first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"User with id {user_id} not found",
-        )
-
-    # Prevent admin from disabling themselves
-    if user.id == current_user.id:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Cannot toggle your own account status",
-        )
-
-    user.is_active = not user.is_active
-    db.commit()
-    db.refresh(user)
-
-    return AdminUserResponse(
-        id=user.id,
-        user_name=user.user_name,
-        email=user.email,
-        role=user.role,
-        auth_source=user.auth_source,
-        is_active=user.is_active,
-        created_at=user.created_at,
-        updated_at=user.updated_at,
-    )
-
-
-@router.put("/users/{user_id}/role", response_model=AdminUserResponse)
-async def update_user_role(
-    role_data: RoleUpdate,
-    user_id: int = Path(..., description="User ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Update user role (admin only)
-    """
-    # Query user directly to avoid decrypt_user_git_info modifying the object
-    user = db.query(User).filter(User.id == user_id).first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"User with id {user_id} not found",
-        )
-
-    # Prevent admin from demoting themselves if they're the only admin
-    if user.id == current_user.id and role_data.role == "user":
-        admin_count = (
-            db.query(User).filter(User.role == "admin", User.is_active == True).count()
-        )
-        if admin_count <= 1:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="Cannot demote yourself when you are the only admin",
-            )
-
-    user.role = role_data.role
-    db.commit()
-    db.refresh(user)
-
-    return AdminUserResponse(
-        id=user.id,
-        user_name=user.user_name,
-        email=user.email,
-        role=user.role,
-        auth_source=user.auth_source,
-        is_active=user.is_active,
-        created_at=user.created_at,
-        updated_at=user.updated_at,
-    )
-
-
-# ==================== Public Model Management Endpoints ====================
-
-
-@router.get("/public-models", response_model=PublicModelListResponse)
-async def list_public_models(
-    page: int = Query(1, ge=1),
-    limit: int = Query(20, ge=1, le=100),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get list of all public models with pagination, sorted by displayName
-    """
-    query = db.query(Kind).filter(
-        Kind.user_id == 0, Kind.kind == "Model", Kind.namespace == "default"
-    )
-    total = query.count()
-    models = query.all()
-
-    # Helper function to extract displayName from json
-    def get_display_name(model: Kind) -> str:
-        """Extract displayName from model json, fallback to name"""
-        if model.json and isinstance(model.json, dict):
-            metadata = model.json.get("metadata", {})
-            if isinstance(metadata, dict):
-                display_name = metadata.get("displayName")
-                if display_name:
-                    return display_name
-        return model.name
-
-    # Sort models by displayName (case-insensitive)
-    sorted_models = sorted(models, key=lambda m: get_display_name(m).lower())
-
-    # Apply pagination after sorting
-    start_idx = (page - 1) * limit
-    end_idx = start_idx + limit
-    paginated_models = sorted_models[start_idx:end_idx]
-
-    return PublicModelListResponse(
-        total=total,
-        items=[
-            PublicModelResponse(
-                id=model.id,
-                name=model.name,
-                namespace=model.namespace,
-                display_name=(
-                    get_display_name(model)
-                    if get_display_name(model) != model.name
-                    else None
-                ),
-                model_json=model.json,
-                is_active=model.is_active,
-                created_at=model.created_at,
-                updated_at=model.updated_at,
-            )
-            for model in paginated_models
-        ],
-    )
-
-
-@router.post(
-    "/public-models",
-    response_model=PublicModelResponse,
-    status_code=status.HTTP_201_CREATED,
-)
-async def create_public_model(
-    model_data: PublicModelCreate,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Create a new public model (admin only)
-    """
-    # Check if model with same name and namespace already exists
-    existing_model = (
-        db.query(Kind)
-        .filter(
-            Kind.user_id == 0,
-            Kind.kind == "Model",
-            Kind.name == model_data.name,
-            Kind.namespace == model_data.namespace,
-        )
-        .first()
-    )
-    if existing_model:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"Public model '{model_data.name}' already exists in namespace '{model_data.namespace}'",
-        )
-
-    new_model = Kind(
-        user_id=0,
-        kind="Model",
-        name=model_data.name,
-        namespace=model_data.namespace,
-        json=model_data.model_json,
-        is_active=True,
-    )
-    db.add(new_model)
-    db.commit()
-    db.refresh(new_model)
-
-    return PublicModelResponse(
-        id=new_model.id,
-        name=new_model.name,
-        namespace=new_model.namespace,
-        model_json=new_model.json,
-        is_active=new_model.is_active,
-        created_at=new_model.created_at,
-        updated_at=new_model.updated_at,
-    )
-
-
-@router.put("/public-models/{model_id}", response_model=PublicModelResponse)
-async def update_public_model(
-    model_data: PublicModelUpdate,
-    model_id: int = Path(..., description="Model ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Update a public model (admin only)
-    """
-    model = (
-        db.query(Kind)
-        .filter(Kind.id == model_id, Kind.user_id == 0, Kind.kind == "Model")
-        .first()
-    )
-    if not model:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Public model with id {model_id} not found",
-        )
-
-    # Check name uniqueness if being changed
-    if model_data.name and model_data.name != model.name:
-        namespace = model_data.namespace or model.namespace
-        existing_model = (
-            db.query(Kind)
-            .filter(
-                Kind.user_id == 0,
-                Kind.kind == "Model",
-                Kind.name == model_data.name,
-                Kind.namespace == namespace,
-            )
-            .first()
-        )
-        if existing_model:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail=f"Public model '{model_data.name}' already exists in namespace '{namespace}'",
-            )
-
-    # Update fields
-    if model_data.name is not None:
-        model.name = model_data.name
-    if model_data.namespace is not None:
-        model.namespace = model_data.namespace
-    if model_data.model_json is not None:
-        model.json = model_data.model_json
-    if model_data.is_active is not None:
-        model.is_active = model_data.is_active
-
-    db.commit()
-    db.refresh(model)
-
-    return PublicModelResponse(
-        id=model.id,
-        name=model.name,
-        namespace=model.namespace,
-        model_json=model.json,
-        is_active=model.is_active,
-        created_at=model.created_at,
-        updated_at=model.updated_at,
-    )
-
-
-@router.delete("/public-models/{model_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_public_model(
-    model_id: int = Path(..., description="Model ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Delete a public model (admin only)
-    """
-    model = (
-        db.query(Kind)
-        .filter(Kind.id == model_id, Kind.user_id == 0, Kind.kind == "Model")
-        .first()
-    )
-    if not model:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Public model with id {model_id} not found",
-        )
-
-    db.delete(model)
-    db.commit()
-
-    return None
-
-
-# ==================== System Stats Endpoint ====================
-
-
-@router.get("/stats", response_model=SystemStats)
-async def get_system_stats(
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get system statistics
-    """
-    from app.models.task import Task
-
-    total_users = db.query(User).count()
-    active_users = db.query(User).filter(User.is_active == True).count()
-    admin_count = (
-        db.query(User).filter(User.role == "admin", User.is_active == True).count()
-    )
-    total_tasks = db.query(Task).count()
-    total_public_models = (
-        db.query(Kind)
-        .filter(Kind.user_id == 0, Kind.kind == "Model", Kind.namespace == "default")
-        .count()
-    )
-
-    return SystemStats(
-        total_users=total_users,
-        active_users=active_users,
-        admin_count=admin_count,
-        total_tasks=total_tasks,
-        total_public_models=total_public_models,
-    )
-
-
-# ==================== Task Management Endpoints ====================
-
-
-@router.post(
-    "/users/{user_id}/tasks",
-    response_model=TaskInDB,
-    status_code=status.HTTP_201_CREATED,
-)
-async def create_task_for_user_id(
-    task: TaskCreate,
-    task_id: Optional[int] = None,
-    user_id: int = Path(..., description="User ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Create task for specified user ID
-    """
-    # Verify user exists
-    target_user = user_service.get_user_by_id(db, user_id)
-
-    # Create task
-    return task_kinds_service.create_task_or_append(
-        db=db, obj_in=task, user=target_user, task_id=task_id
-    )
-
-
-@router.post(
-    "/users/username/{user_name}/tasks",
-    response_model=TaskInDB,
-    status_code=status.HTTP_201_CREATED,
-)
-async def create_task_for_user_by_username(
-    task: TaskCreate,
-    task_id: Optional[int] = None,
-    user_name: str = Path(..., description="User name"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Create task for specified user name
-    """
-    # Verify user exists
-    target_user = user_service.get_user_by_name(db, user_name)
-
-    # Create task
-    return task_kinds_service.create_task_or_append(
-        db=db, obj_in=task, user=target_user, task_id=task_id
-    )
-
-
-@router.post("/generate-admin-token", response_model=Token)
-async def generate_admin_token(
-    db: Session = Depends(get_db), current_user: User = Depends(get_admin_user)
-):
-    """
-    Generate a permanent admin token (pseudo-permanent for 500 years)
-    """
-    # Create a permanent token (set very long expiration time)
-    access_token = create_access_token(
-        data={"sub": current_user.user_name, "user_id": current_user.id},
-        expires_delta=262800000,  # 500 years
-    )
-
-    return Token(access_token=access_token, token_type="bearer")
-
-
-# Admin Kind Management Endpoints
-# Provide administrators with full access to all user resources
-
-
-@router.get("/users/{user_id}/kinds/{kinds}")
-async def admin_list_user_resources(
-    user_id: int = Path(..., description="User ID"),
-    kinds: str = Path(
-        ...,
-        description="Resource type. Valid options: ghosts, models, shells, bots, teams, workspaces, tasks",
-    ),
-    namespace: str = Query("default", description="Resource namespace"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get all resources of specified type for a user
-
-    Administrators can view resource lists for any user.
-    """
-    # Validate resource type
-    kind = validate_resource_type(kinds)
-
-    # Verify user exists
-    validate_user_exists(db, user_id)
-
-    # Get resource list
-    resources = kind_service.list_resources(user_id, kind, namespace)
-
-    # Format and return response
-    return format_resource_list(kind, resources)
-
-
-@router.get("/users/username/{user_name}/kinds/{kinds}")
-async def admin_list_user_resources_by_username(
-    user_name: str = Path(..., description="User name"),
-    kinds: str = Path(
-        ...,
-        description="Resource type. Valid options: ghosts, models, shells, bots, teams, workspaces, tasks",
-    ),
-    namespace: str = Query("default", description="Resource namespace"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get all resources of specified type for a user by username
-
-    Administrators can view resource lists for any user.
-    """
-    # Validate resource type
-    kind = validate_resource_type(kinds)
-
-    # Verify user exists and get user ID
-    target_user = user_service.get_user_by_name(db, user_name)
-    user_id = target_user.id
-
-    # Get resource list
-    resources = kind_service.list_resources(user_id, kind, namespace)
-
-    # Format and return response
-    return format_resource_list(kind, resources)
-
-
-@router.get("/users/{user_id}/kinds/{kinds}/{name}")
-async def admin_get_user_resource(
-    user_id: int = Path(..., description="User ID"),
-    kinds: str = Path(
-        ...,
-        description="Resource type. Valid options: ghosts, models, shells, bots, teams, workspaces, tasks",
-    ),
-    name: str = Path(..., description="Resource name"),
-    namespace: str = Query("default", description="Resource namespace"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get specific resource for a user
-
-    Administrators can view details of any specific resource for any user.
-    """
-    # Validate resource type
-    kind = validate_resource_type(kinds)
-
-    # Verify user exists
-    validate_user_exists(db, user_id)
-
-    # Get resource
-    resource = kind_service.get_resource(user_id, kind, namespace, name)
-    if not resource:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"{kind} resource '{name}' not found in namespace '{namespace}'",
-        )
-
-    # Format and return response
-    return format_single_resource(kind, resource)
-
-
-@router.get("/users/username/{user_name}/kinds/{kinds}/{name}")
-async def admin_get_user_resource_by_username(
-    user_name: str = Path(..., description="User name"),
-    kinds: str = Path(
-        ...,
-        description="Resource type. Valid options: ghosts, models, shells, bots, teams, workspaces, tasks",
-    ),
-    name: str = Path(..., description="Resource name"),
-    namespace: str = Query("default", description="Resource namespace"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get specific resource for a user by username
-
-    Administrators can view details of any specific resource for any user.
-    """
-    # Validate resource type
-    kind = validate_resource_type(kinds)
-
-    # Verify user exists and get user ID
-    target_user = user_service.get_user_by_name(db, user_name)
-    user_id = target_user.id
-
-    # Get resource
-    resource = kind_service.get_resource(user_id, kind, namespace, name)
-    if not resource:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"{kind} resource '{name}' not found in namespace '{namespace}'",
-        )
-
-    # Format and return response
-    return format_single_resource(kind, resource)
-
-
-@router.post("/users/{user_id}/kinds/{kinds}", status_code=status.HTTP_201_CREATED)
-async def admin_create_resource_for_user(
-    user_id: int = Path(..., description="User ID"),
-    kinds: str = Path(
-        ...,
-        description="Resource type. Valid options: ghosts, models, shells, bots, teams, workspaces, tasks",
-    ),
-    namespace: str = Query("default", description="Resource namespace"),
-    resource: Dict[str, Any] = None,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Create resource for specified user
-
-    Administrators can create resources for any user.
-    """
-    # Validate resource type
-    kind = validate_resource_type(kinds)
-
-    # Verify user exists
-    validate_user_exists(db, user_id)
-
-    # Validate and prepare resource data
-    validated_resource = validate_and_prepare_resource(kind, resource, namespace)
-
-    # Create resource
-    resource_id = kind_service.create_resource(user_id, kind, validated_resource)
-
-    # Format and return response
-    formatted_resource = kind_service._format_resource_by_id(kind, resource_id)
-    schema_class = KIND_SCHEMA_MAP[kind]
-    return schema_class.parse_obj(formatted_resource)
-
-
-@router.put("/users/{user_id}/kinds/{kinds}/{name}")
-async def admin_update_user_resource(
-    user_id: int = Path(..., description="User ID"),
-    kinds: str = Path(
-        ...,
-        description="Resource type. Valid options: ghosts, models, shells, bots, teams, workspaces, tasks",
-    ),
-    name: str = Path(..., description="Resource name"),
-    namespace: str = Query("default", description="Resource namespace"),
-    resource: Dict[str, Any] = None,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Update resource for specified user
-
-    Administrators can update resources for any user.
-    """
-    # Validate resource type
-    kind = validate_resource_type(kinds)
-
-    # Verify user exists
-    validate_user_exists(db, user_id)
-
-    # Validate and prepare resource data
-    validated_resource = validate_and_prepare_resource(kind, resource, namespace, name)
-
-    # Update resource
-    resource_id = kind_service.update_resource(
-        user_id, kind, namespace, name, validated_resource
-    )
-
-    # Format and return response
-    formatted_resource = kind_service._format_resource_by_id(kind, resource_id)
-    schema_class = KIND_SCHEMA_MAP[kind]
-    return schema_class.parse_obj(formatted_resource)
-
-
-@router.delete(
-    "/users/{user_id}/kinds/{kinds}/{name}", status_code=status.HTTP_204_NO_CONTENT
-)
-async def admin_delete_user_resource(
-    user_id: int = Path(..., description="User ID"),
-    kinds: str = Path(
-        ...,
-        description="Resource type. Valid options: ghosts, models, shells, bots, teams, workspaces, tasks",
-    ),
-    name: str = Path(..., description="Resource name"),
-    namespace: str = Query("default", description="Resource namespace"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Delete resource for specified user
-
-    Administrators can delete resources for any user.
-    """
-    # Validate resource type
-    kind = validate_resource_type(kinds)
-
-    # Verify user exists
-    validate_user_exists(db, user_id)
-
-    # Delete resource
-    kind_service.delete_resource(user_id, kind, namespace, name)
-
-    return {
-        "message": f"Successfully deleted {kind} resource '{name}' for user {user_id}"
-    }
-
-
-# Admin Batch Operation Endpoints
-# Provide administrators with batch operation capabilities for user resources
-
-
-@router.post("/users/{user_id}/kinds/batch/apply", response_model=BatchResponse)
-async def admin_apply_resources_for_user(
-    user_id: int = Path(..., description="User ID"),
-    namespace: str = Query("default", description="Resource namespace"),
-    resources: List[Dict[str, Any]] = None,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Batch apply resources for specified user (create or update)
-
-    Administrators can batch create or update resources for any user.
-    """
-    if not resources:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST, detail="Resource list is required"
-        )
-
-    # Verify user exists
-    user_service.get_user_by_id(db, user_id)
-
-    # Ensure all resources have correct namespace
-    for resource in resources:
-        if "metadata" not in resource:
-            resource["metadata"] = {}
-        resource["metadata"]["namespace"] = namespace
-
-    try:
-        # Execute batch operation
-        results = batch_service.apply_resources(user_id, resources)
-
-        success_count = sum(1 for r in results if r["success"])
-        total_count = len(results)
-
-        return BatchResponse(
-            success=success_count == total_count,
-            message=f"Applied {success_count}/{total_count} resources for user {user_id}",
-            results=results,
-        )
-    except Exception as e:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"Error batch applying resources for user {user_id}: {str(e)}",
-        )
-
-
-@router.post("/users/{user_id}/kinds/batch/delete", response_model=BatchResponse)
-async def admin_delete_resources_for_user(
-    user_id: int = Path(..., description="User ID"),
-    namespace: str = Query("default", description="Resource namespace"),
-    resources: List[Dict[str, Any]] = None,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Batch delete resources for specified user
-
-    Administrators can batch delete resources for any user.
-    """
-    # Verify user exists
-    validate_user_exists(db, user_id)
-
-    # Prepare batch resource data
-    prepare_batch_resources(resources, namespace)
-
-    # Execute batch delete operation
-    results = batch_service.delete_resources(user_id, resources)
-
-    success_count = sum(1 for r in results if r["success"])
-    total_count = len(results)
-
-    return BatchResponse(
-        success=success_count == total_count,
-        message=f"Deleted {success_count}/{total_count} resources for user {user_id}",
-        results=results,
-    )
-
-
-# ==================== System Config Endpoints ====================
-
-QUICK_ACCESS_CONFIG_KEY = "quick_access_recommended"
-
-
-@router.get("/system-config/quick-access", response_model=SystemConfigResponse)
-async def get_quick_access_config(
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get system recommended quick access configuration
-    """
-    config = (
-        db.query(SystemConfig)
-        .filter(SystemConfig.config_key == QUICK_ACCESS_CONFIG_KEY)
-        .first()
-    )
-    if not config:
-        return SystemConfigResponse(version=0, teams=[])
-
-    config_value = config.config_value or {}
-    return SystemConfigResponse(
-        version=config.version,
-        teams=config_value.get("teams", []),
-    )
-
-
-@router.put("/system-config/quick-access", response_model=SystemConfigResponse)
-async def update_quick_access_config(
-    config_data: SystemConfigUpdate,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Update system recommended quick access configuration (admin only).
-    Version number is automatically incremented.
-    """
-    config = (
-        db.query(SystemConfig)
-        .filter(SystemConfig.config_key == QUICK_ACCESS_CONFIG_KEY)
-        .first()
-    )
-
-    if not config:
-        # Create new config
-        config = SystemConfig(
-            config_key=QUICK_ACCESS_CONFIG_KEY,
-            config_value={"teams": config_data.teams},
-            version=1,
-            updated_by=current_user.id,
-        )
-        db.add(config)
-    else:
-        # Update existing config and increment version
-        config.config_value = {"teams": config_data.teams}
-        config.version = config.version + 1
-        config.updated_by = current_user.id
-
-    db.commit()
-    db.refresh(config)
-
-    return SystemConfigResponse(
-        version=config.version,
-        teams=config.config_value.get("teams", []),
-    )
-
-
-# ==================== Chat Slogan & Tips Config Endpoints ====================
-
-CHAT_SLOGAN_TIPS_CONFIG_KEY = "chat_slogan_tips"
-
-# Default slogan and tips configuration
-DEFAULT_SLOGAN_TIPS_CONFIG = {
-    "slogans": [
-        {
-            "id": 1,
-            "zh": "今天有什么可以帮到你？",
-            "en": "What can I help you with today?",
-            "mode": "chat",
-        },
-        {
-            "id": 2,
-            "zh": "让我们一起写代码吧",
-            "en": "Let's code together",
-            "mode": "code",
-        },
-    ],
-    "tips": [
-        {
-            "id": 1,
-            "zh": "试试问我：帮我分析这段代码的性能问题",
-            "en": "Try asking: Help me analyze the performance issues in this code",
-        },
-        {
-            "id": 2,
-            "zh": "你可以上传文件让我帮你处理",
-            "en": "You can upload files for me to help you process",
-        },
-        {
-            "id": 3,
-            "zh": "我可以帮你生成代码、修复 Bug 或重构现有代码",
-            "en": "I can help you generate code, fix bugs, or refactor existing code",
-        },
-        {
-            "id": 4,
-            "zh": "试试让我帮你编写单元测试或文档",
-            "en": "Try asking me to write unit tests or documentation",
-        },
-        {
-            "id": 5,
-            "zh": "我可以解释复杂的代码逻辑，帮助你理解代码库",
-            "en": "I can explain complex code logic and help you understand the codebase",
-        },
-    ],
-}
-
-
-@router.get("/system-config/slogan-tips", response_model=ChatSloganTipsResponse)
-async def get_slogan_tips_config(
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get chat slogan and tips configuration
-    """
-    config = (
-        db.query(SystemConfig)
-        .filter(SystemConfig.config_key == CHAT_SLOGAN_TIPS_CONFIG_KEY)
-        .first()
-    )
-    if not config:
-        # Return default configuration
-        return ChatSloganTipsResponse(
-            version=0,
-            slogans=[
-                ChatSloganItem(**s) for s in DEFAULT_SLOGAN_TIPS_CONFIG["slogans"]
-            ],
-            tips=[ChatTipItem(**tip) for tip in DEFAULT_SLOGAN_TIPS_CONFIG["tips"]],
-        )
-
-    config_value = config.config_value or {}
-    return ChatSloganTipsResponse(
-        version=config.version,
-        slogans=[
-            ChatSloganItem(**s)
-            for s in config_value.get("slogans", DEFAULT_SLOGAN_TIPS_CONFIG["slogans"])
-        ],
-        tips=[ChatTipItem(**tip) for tip in config_value.get("tips", [])],
-    )
-
-
-@router.put("/system-config/slogan-tips", response_model=ChatSloganTipsResponse)
-async def update_slogan_tips_config(
-    config_data: ChatSloganTipsUpdate,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Update chat slogan and tips configuration (admin only).
-    Version number is automatically incremented.
-    """
-    config = (
-        db.query(SystemConfig)
-        .filter(SystemConfig.config_key == CHAT_SLOGAN_TIPS_CONFIG_KEY)
-        .first()
-    )
-
-    config_value = {
-        "slogans": [s.model_dump() for s in config_data.slogans],
-        "tips": [tip.model_dump() for tip in config_data.tips],
-    }
-
-    if not config:
-        # Create new config
-        config = SystemConfig(
-            config_key=CHAT_SLOGAN_TIPS_CONFIG_KEY,
-            config_value=config_value,
-            version=1,
-            updated_by=current_user.id,
-        )
-        db.add(config)
-    else:
-        # Update existing config and increment version
-        config.config_value = config_value
-        config.version = config.version + 1
-        config.updated_by = current_user.id
-
-    db.commit()
-    db.refresh(config)
-
-    return ChatSloganTipsResponse(
-        version=config.version,
-        slogans=config_data.slogans,
-        tips=config_data.tips,
-    )
-
-
-# ==================== Service Key Management Endpoints ====================
-
-import hashlib
-import secrets
-
-from app.models.api_key import KEY_TYPE_PERSONAL, KEY_TYPE_SERVICE, APIKey
-from app.schemas.api_key import (
-    AdminPersonalKeyListResponse,
-    AdminPersonalKeyResponse,
-    ServiceKeyCreate,
-    ServiceKeyCreatedResponse,
-    ServiceKeyListResponse,
-    ServiceKeyResponse,
-)
-
-
-@router.get("/service-keys", response_model=ServiceKeyListResponse)
-async def list_service_keys(
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get list of all service keys (admin only), including disabled ones.
-
-    Service keys are used for trusted service authentication.
-    """
-    # Query service keys with creator information
-    results = (
-        db.query(APIKey, User)
-        .outerjoin(User, APIKey.user_id == User.id)
-        .filter(
-            APIKey.key_type == KEY_TYPE_SERVICE,
-        )
-        .order_by(APIKey.created_at.desc())
-        .all()
-    )
-
-    items = []
-    for api_key, creator in results:
-        items.append(
-            ServiceKeyResponse(
-                id=api_key.id,
-                name=api_key.name,
-                key_prefix=api_key.key_prefix,
-                description=api_key.description,
-                expires_at=api_key.expires_at,
-                last_used_at=api_key.last_used_at,
-                created_at=api_key.created_at,
-                is_active=api_key.is_active,
-                created_by=creator.user_name if creator else None,
-            )
-        )
-
-    return ServiceKeyListResponse(items=items, total=len(items))
-
-
-@router.post(
-    "/service-keys",
-    response_model=ServiceKeyCreatedResponse,
-    status_code=status.HTTP_201_CREATED,
-)
-async def create_service_key(
-    service_key_create: ServiceKeyCreate,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Create a new service key (admin only).
-
-    The full key is only returned once at creation time.
-    Store it securely as it cannot be retrieved again.
-
-    Service keys are used for trusted service authentication
-    via the wegent-source header.
-    """
-    # Generate key: wg-{32 random chars}
-    random_part = secrets.token_urlsafe(32)
-    full_key = f"wg-{random_part}"
-
-    # Hash the key for storage
-    key_hash = hashlib.sha256(full_key.encode()).hexdigest()
-
-    # Create prefix for display (first 8 chars after "wg-")
-    key_prefix = f"wg-{random_part[:8]}..."
-
-    # Create the service key record (user_id records the creator admin)
-    service_key = APIKey(
-        user_id=current_user.id,
-        key_hash=key_hash,
-        key_prefix=key_prefix,
-        name=service_key_create.name,
-        key_type=KEY_TYPE_SERVICE,
-        description=service_key_create.description or "",
-    )
-
-    db.add(service_key)
-    db.commit()
-    db.refresh(service_key)
-
-    # Return with full key (only shown once)
-    return ServiceKeyCreatedResponse(
-        id=service_key.id,
-        name=service_key.name,
-        key_prefix=service_key.key_prefix,
-        description=service_key.description,
-        key=full_key,
-        expires_at=service_key.expires_at,
-        last_used_at=service_key.last_used_at,
-        created_at=service_key.created_at,
-        is_active=service_key.is_active,
-        created_by=current_user.user_name,
-    )
-
-
-@router.post("/service-keys/{key_id}/toggle-status", response_model=ServiceKeyResponse)
-async def toggle_service_key_status(
-    key_id: int,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Toggle a service key's active status (admin only).
-
-    Enable or disable a service key without deleting it.
-    """
-    result = (
-        db.query(APIKey, User)
-        .outerjoin(User, APIKey.user_id == User.id)
-        .filter(
-            APIKey.id == key_id,
-            APIKey.key_type == KEY_TYPE_SERVICE,
-        )
-        .first()
-    )
-
-    if not result:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Service key not found",
-        )
-
-    service_key, creator = result
-
-    # Toggle is_active status
-    service_key.is_active = not service_key.is_active
-    db.commit()
-    db.refresh(service_key)
-
-    return ServiceKeyResponse(
-        id=service_key.id,
-        name=service_key.name,
-        key_prefix=service_key.key_prefix,
-        description=service_key.description,
-        expires_at=service_key.expires_at,
-        last_used_at=service_key.last_used_at,
-        created_at=service_key.created_at,
-        is_active=service_key.is_active,
-        created_by=creator.user_name if creator else None,
-    )
-
-
-@router.delete("/service-keys/{key_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_service_key(
-    key_id: int,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Delete a service key (admin only).
-
-    This is a hard delete - the key will be permanently removed.
-    """
-    service_key = (
-        db.query(APIKey)
-        .filter(
-            APIKey.id == key_id,
-            APIKey.key_type == KEY_TYPE_SERVICE,
-        )
-        .first()
-    )
-
-    if not service_key:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Service key not found",
-        )
-
-    # Hard delete - permanently remove the record
-    db.delete(service_key)
-    db.commit()
-
-    return None
-
-
-# ==================== Personal Key Management Endpoints (Admin) ====================
-
-
-@router.get("/personal-keys", response_model=AdminPersonalKeyListResponse)
-async def list_all_personal_keys(
-    page: int = Query(1, ge=1),
-    limit: int = Query(50, ge=1, le=100),
-    search: Optional[str] = Query(None, description="Search by username or key name"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get list of all personal keys with their owners (admin only).
-
-    Personal keys are user-created API keys for programmatic access.
-    """
-    query = (
-        db.query(APIKey, User)
-        .join(User, APIKey.user_id == User.id)
-        .filter(APIKey.key_type == KEY_TYPE_PERSONAL)
-    )
-
-    # Apply search filter
-    if search:
-        search_pattern = f"%{search}%"
-        query = query.filter(
-            (User.user_name.ilike(search_pattern)) | (APIKey.name.ilike(search_pattern))
-        )
-
-    total = query.count()
-    results = (
-        query.order_by(APIKey.created_at.desc())
-        .offset((page - 1) * limit)
-        .limit(limit)
-        .all()
-    )
-
-    items = []
-    for api_key, user in results:
-        items.append(
-            AdminPersonalKeyResponse(
-                id=api_key.id,
-                user_id=api_key.user_id,
-                user_name=user.user_name,
-                name=api_key.name,
-                key_prefix=api_key.key_prefix,
-                description=api_key.description,
-                expires_at=api_key.expires_at,
-                last_used_at=api_key.last_used_at,
-                created_at=api_key.created_at,
-                is_active=api_key.is_active,
-            )
-        )
-
-    return AdminPersonalKeyListResponse(items=items, total=total)
-
-
-@router.post(
-    "/personal-keys/{key_id}/toggle-status", response_model=AdminPersonalKeyResponse
-)
-async def toggle_personal_key_status(
-    key_id: int,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Toggle a personal key's active status (admin only).
-
-    Enable or disable a personal key without deleting it.
-    """
-    result = (
-        db.query(APIKey, User)
-        .join(User, APIKey.user_id == User.id)
-        .filter(
-            APIKey.id == key_id,
-            APIKey.key_type == KEY_TYPE_PERSONAL,
-        )
-        .first()
-    )
-
-    if not result:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Personal key not found",
-        )
-
-    api_key, user = result
-
-    # Toggle is_active status
-    api_key.is_active = not api_key.is_active
-    db.commit()
-    db.refresh(api_key)
-
-    return AdminPersonalKeyResponse(
-        id=api_key.id,
-        user_id=api_key.user_id,
-        user_name=user.user_name,
-        name=api_key.name,
-        key_prefix=api_key.key_prefix,
-        description=api_key.description,
-        expires_at=api_key.expires_at,
-        last_used_at=api_key.last_used_at,
-        created_at=api_key.created_at,
-        is_active=api_key.is_active,
-    )
-
-
-@router.delete("/personal-keys/{key_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_personal_key(
-    key_id: int,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Delete a personal key (admin only).
-
-    This is a hard delete - the key will be permanently removed.
-    """
-    api_key = (
-        db.query(APIKey)
-        .filter(
-            APIKey.id == key_id,
-            APIKey.key_type == KEY_TYPE_PERSONAL,
-        )
-        .first()
-    )
-
-    if not api_key:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Personal key not found",
-        )
-
-    # Hard delete - permanently remove the record
-    db.delete(api_key)
-    db.commit()
-
-    return None
-
-
-# ==================== Public Retriever Management Endpoints ====================
-
-
-@router.get("/public-retrievers", response_model=PublicRetrieverListResponse)
-async def list_public_retrievers(
-    page: int = Query(1, ge=1),
-    limit: int = Query(20, ge=1, le=100),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get list of all public retrievers with pagination
-    """
-    total = public_retriever_service.count_active_retrievers(
-        db, current_user=current_user
-    )
-    skip = (page - 1) * limit
-    retrievers = public_retriever_service.get_retrievers(
-        db, skip=skip, limit=limit, current_user=current_user
-    )
-
-    return PublicRetrieverListResponse(
-        total=total,
-        items=[
-            PublicRetrieverResponse(
-                id=r["id"],
-                name=r["name"],
-                namespace=r["namespace"],
-                displayName=r["displayName"],
-                storageType=r["storageType"],
-                description=r["description"],
-                json=r["json"],
-                is_active=r["is_active"],
-                created_at=r["created_at"],
-                updated_at=r["updated_at"],
-            )
-            for r in retrievers
-        ],
-    )
-
-
-@router.post(
-    "/public-retrievers",
-    response_model=PublicRetrieverResponse,
-    status_code=status.HTTP_201_CREATED,
-)
-async def create_public_retriever(
-    retriever_data: Retriever,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Create a new public retriever (admin only)
-    """
-    result = public_retriever_service.create_retriever(
-        db, retriever=retriever_data, current_user=current_user
-    )
-    return PublicRetrieverResponse(
-        id=result["id"],
-        name=result["name"],
-        namespace=result["namespace"],
-        displayName=result["displayName"],
-        storageType=result["storageType"],
-        description=result["description"],
-        json=result["json"],
-        is_active=result["is_active"],
-        created_at=result["created_at"],
-        updated_at=result["updated_at"],
-    )
-
-
-@router.put("/public-retrievers/{retriever_id}", response_model=PublicRetrieverResponse)
-async def update_public_retriever(
-    retriever_data: Retriever,
-    retriever_id: int = Path(..., description="Retriever ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Update a public retriever (admin only)
-    """
-    result = public_retriever_service.update_retriever(
-        db,
-        retriever_id=retriever_id,
-        retriever=retriever_data,
-        current_user=current_user,
-    )
-    return PublicRetrieverResponse(
-        id=result["id"],
-        name=result["name"],
-        namespace=result["namespace"],
-        displayName=result["displayName"],
-        storageType=result["storageType"],
-        description=result["description"],
-        json=result["json"],
-        is_active=result["is_active"],
-        created_at=result["created_at"],
-        updated_at=result["updated_at"],
-    )
-
-
-@router.delete(
-    "/public-retrievers/{retriever_id}", status_code=status.HTTP_204_NO_CONTENT
-)
-async def delete_public_retriever(
-    retriever_id: int = Path(..., description="Retriever ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Delete a public retriever (admin only)
-    """
-    public_retriever_service.delete_retriever(
-        db, retriever_id=retriever_id, current_user=current_user
-    )
-    return None
-
-
-# ==================== Public Team Management Endpoints ====================
-
-
-def _get_team_description(team: Kind) -> Optional[str]:
-    """Extract description from team json"""
-    if team.json and isinstance(team.json, dict):
-        spec = team.json.get("spec", {})
-        if isinstance(spec, dict):
-            return spec.get("description")
-    return None
-
-
-def _get_team_display_name(team: Kind) -> Optional[str]:
-    """Extract displayName from team json metadata"""
-    if team.json and isinstance(team.json, dict):
-        metadata = team.json.get("metadata", {})
-        if isinstance(metadata, dict):
-            display_name = metadata.get("displayName")
-            if display_name and display_name != team.name:
-                return display_name
-    return None
-
-
-def _validate_team_bot_references(
-    db: Session, team_json: dict
-) -> tuple[bool, Optional[str]]:
-    """
-    Validate that all bots referenced by the team are public bots.
-
-    Returns:
-        (is_valid, error_message)
-    """
-    # Defensive type checking
-    if not isinstance(team_json, dict):
-        return (False, "Invalid team JSON: must be an object")
-
-    spec = team_json.get("spec", {})
-    if not isinstance(spec, dict):
-        return (False, "Invalid team JSON: 'spec' must be an object")
-
-    members = spec.get("members", [])
-    if not isinstance(members, list):
-        return (False, "Invalid team JSON: 'spec.members' must be an array")
-
-    for member in members:
-        if not isinstance(member, dict):
-            continue
-        bot_ref = member.get("botRef", {})
-        if not isinstance(bot_ref, dict):
-            continue
-        bot_name = bot_ref.get("name")
-        if not isinstance(bot_name, str) or not bot_name:
-            continue
-        bot_namespace = bot_ref.get("namespace", "default")
-
-        # Check if bot exists as a public resource (user_id=0)
-        bot = (
-            db.query(Kind)
-            .filter(
-                Kind.user_id == 0,
-                Kind.kind == "Bot",
-                Kind.name == bot_name,
-                Kind.namespace == bot_namespace,
-                Kind.is_active == True,
-            )
-            .first()
-        )
-
-        if not bot:
-            return (
-                False,
-                f"Bot '{bot_namespace}/{bot_name}' is not a public resource. Please create it as a public bot first.",
-            )
-
-    return (True, None)
-
-
-@router.get("/public-teams", response_model=PublicTeamListResponse)
-async def list_public_teams(
-    page: int = Query(1, ge=1),
-    limit: int = Query(20, ge=1, le=100),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get list of all public teams with pagination
-    """
-    query = db.query(Kind).filter(Kind.user_id == 0, Kind.kind == "Team")
-    total = query.count()
-
-    # Apply SQL-level pagination
-    paginated_teams = (
-        query.order_by(Kind.updated_at.desc())
-        .offset((page - 1) * limit)
-        .limit(limit)
-        .all()
-    )
-
-    return PublicTeamListResponse(
-        total=total,
-        items=[
-            PublicTeamResponse(
-                id=team.id,
-                name=team.name,
-                namespace=team.namespace,
-                display_name=_get_team_display_name(team),
-                description=_get_team_description(team),
-                team_json=team.json,
-                is_active=team.is_active,
-                created_at=team.created_at,
-                updated_at=team.updated_at,
-            )
-            for team in paginated_teams
-        ],
-    )
-
-
-@router.post(
-    "/public-teams",
-    response_model=PublicTeamResponse,
-    status_code=status.HTTP_201_CREATED,
-)
-async def create_public_team(
-    team_data: PublicTeamCreate,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Create a new public team (admin only).
-    All bots referenced by the team must be public bots (user_id=0).
-    """
-    # Check if team with same name and namespace already exists
-    existing_team = (
-        db.query(Kind)
-        .filter(
-            Kind.user_id == 0,
-            Kind.kind == "Team",
-            Kind.name == team_data.name,
-            Kind.namespace == team_data.namespace,
-        )
-        .first()
-    )
-    if existing_team:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"Public team '{team_data.name}' already exists in namespace '{team_data.namespace}'",
-        )
-
-    # Validate bot references
-    is_valid, error_message = _validate_team_bot_references(db, team_data.team_json)
-    if not is_valid:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=error_message,
-        )
-
-    new_team = Kind(
-        user_id=0,
-        kind="Team",
-        name=team_data.name,
-        namespace=team_data.namespace,
-        json=team_data.team_json,
-        is_active=True,
-    )
-    db.add(new_team)
-    db.commit()
-    db.refresh(new_team)
-
-    return PublicTeamResponse(
-        id=new_team.id,
-        name=new_team.name,
-        namespace=new_team.namespace,
-        display_name=_get_team_display_name(new_team),
-        description=_get_team_description(new_team),
-        team_json=new_team.json,
-        is_active=new_team.is_active,
-        created_at=new_team.created_at,
-        updated_at=new_team.updated_at,
-    )
-
-
-@router.put("/public-teams/{team_id}", response_model=PublicTeamResponse)
-async def update_public_team(
-    team_data: PublicTeamUpdate,
-    team_id: int = Path(..., description="Team ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Update a public team (admin only).
-    All bots referenced by the team must be public bots (user_id=0).
-    """
-    team = (
-        db.query(Kind)
-        .filter(Kind.id == team_id, Kind.user_id == 0, Kind.kind == "Team")
-        .first()
-    )
-    if not team:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Public team with id {team_id} not found",
-        )
-
-    # Check name uniqueness if being changed
-    if team_data.name and team_data.name != team.name:
-        namespace = team_data.namespace or team.namespace
-        existing_team = (
-            db.query(Kind)
-            .filter(
-                Kind.user_id == 0,
-                Kind.kind == "Team",
-                Kind.name == team_data.name,
-                Kind.namespace == namespace,
-            )
-            .first()
-        )
-        if existing_team:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail=f"Public team '{team_data.name}' already exists in namespace '{namespace}'",
-            )
-
-    # Validate bot references if json is being updated
-    if team_data.team_json:
-        is_valid, error_message = _validate_team_bot_references(db, team_data.team_json)
-        if not is_valid:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail=error_message,
-            )
-
-    # Update fields
-    if team_data.name is not None:
-        team.name = team_data.name
-    if team_data.namespace is not None:
-        team.namespace = team_data.namespace
-    if team_data.team_json is not None:
-        team.json = team_data.team_json
-    if team_data.is_active is not None:
-        team.is_active = team_data.is_active
-
-    db.commit()
-    db.refresh(team)
-
-    return PublicTeamResponse(
-        id=team.id,
-        name=team.name,
-        namespace=team.namespace,
-        display_name=_get_team_display_name(team),
-        description=_get_team_description(team),
-        team_json=team.json,
-        is_active=team.is_active,
-        created_at=team.created_at,
-        updated_at=team.updated_at,
-    )
-
-
-@router.delete("/public-teams/{team_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_public_team(
-    team_id: int = Path(..., description="Team ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Delete a public team (admin only)
-    """
-    team = (
-        db.query(Kind)
-        .filter(Kind.id == team_id, Kind.user_id == 0, Kind.kind == "Team")
-        .first()
-    )
-    if not team:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Public team with id {team_id} not found",
-        )
-
-    db.delete(team)
-    db.commit()
-
-    return None
-
-
-# ==================== Public Bot Management Endpoints ====================
-
-
-def _get_bot_ref_info(
-    bot: Kind,
-) -> tuple[Optional[str], Optional[str], Optional[str]]:
-    """Extract ghost, shell, and model reference names from bot json"""
-    ghost_name = None
-    shell_name = None
-    model_name = None
-
-    if bot.json and isinstance(bot.json, dict):
-        spec = bot.json.get("spec", {})
-        if isinstance(spec, dict):
-            ghost_ref = spec.get("ghostRef", {})
-            shell_ref = spec.get("shellRef", {})
-            model_ref = spec.get("modelRef", {})
-
-            if isinstance(ghost_ref, dict):
-                ghost_name = ghost_ref.get("name")
-            if isinstance(shell_ref, dict):
-                shell_name = shell_ref.get("name")
-            if isinstance(model_ref, dict):
-                model_name = model_ref.get("name")
-
-    return (ghost_name, shell_name, model_name)
-
-
-def _get_bot_display_name(bot: Kind) -> Optional[str]:
-    """Extract displayName from bot json metadata"""
-    if bot.json and isinstance(bot.json, dict):
-        metadata = bot.json.get("metadata", {})
-        if isinstance(metadata, dict):
-            display_name = metadata.get("displayName")
-            if display_name and display_name != bot.name:
-                return display_name
-    return None
-
-
-def _validate_bot_resource_references(
-    db: Session, bot_json: dict
-) -> tuple[bool, Optional[str]]:
-    """
-    Validate that all resources referenced by the bot (Ghost, Shell, Model) are public.
-
-    Returns:
-        (is_valid, error_message)
-    """
-    # Defensive type checking
-    if not isinstance(bot_json, dict):
-        return (False, "Invalid bot JSON: must be an object")
-
-    spec = bot_json.get("spec", {})
-    if not isinstance(spec, dict):
-        return (False, "Invalid bot JSON: 'spec' must be an object")
-
-    # Validate ghostRef
-    ghost_ref = spec.get("ghostRef", {})
-    if ghost_ref and not isinstance(ghost_ref, dict):
-        return (False, "Invalid bot JSON: 'spec.ghostRef' must be an object")
-    if isinstance(ghost_ref, dict):
-        ghost_name = ghost_ref.get("name")
-        ghost_namespace = ghost_ref.get("namespace", "default")
-        if ghost_name and isinstance(ghost_name, str):
-            ghost = (
-                db.query(Kind)
-                .filter(
-                    Kind.user_id == 0,
-                    Kind.kind == "Ghost",
-                    Kind.name == ghost_name,
-                    Kind.namespace == ghost_namespace,
-                    Kind.is_active == True,
-                )
-                .first()
-            )
-            if not ghost:
-                return (
-                    False,
-                    f"Ghost '{ghost_namespace}/{ghost_name}' is not a public resource. Please create it as a public ghost first.",
-                )
-
-    # Validate shellRef
-    shell_ref = spec.get("shellRef", {})
-    if shell_ref and not isinstance(shell_ref, dict):
-        return (False, "Invalid bot JSON: 'spec.shellRef' must be an object")
-    if isinstance(shell_ref, dict):
-        shell_name = shell_ref.get("name")
-        shell_namespace = shell_ref.get("namespace", "default")
-        if shell_name and isinstance(shell_name, str):
-            shell = (
-                db.query(Kind)
-                .filter(
-                    Kind.user_id == 0,
-                    Kind.kind == "Shell",
-                    Kind.name == shell_name,
-                    Kind.namespace == shell_namespace,
-                    Kind.is_active == True,
-                )
-                .first()
-            )
-            if not shell:
-                return (
-                    False,
-                    f"Shell '{shell_namespace}/{shell_name}' is not a public resource. Please create it as a public shell first.",
-                )
-
-    # Validate modelRef (optional)
-    model_ref = spec.get("modelRef")
-    if model_ref:
-        if not isinstance(model_ref, dict):
-            return (False, "Invalid bot JSON: 'spec.modelRef' must be an object")
-        model_name = model_ref.get("name")
-        model_namespace = model_ref.get("namespace", "default")
-        if model_name and isinstance(model_name, str):
-            model = (
-                db.query(Kind)
-                .filter(
-                    Kind.user_id == 0,
-                    Kind.kind == "Model",
-                    Kind.name == model_name,
-                    Kind.namespace == model_namespace,
-                    Kind.is_active == True,
-                )
-                .first()
-            )
-            if not model:
-                return (
-                    False,
-                    f"Model '{model_namespace}/{model_name}' is not a public resource. Please create it as a public model first.",
-                )
-
-    return (True, None)
-
-
-@router.get("/public-bots", response_model=PublicBotListResponse)
-async def list_public_bots(
-    page: int = Query(1, ge=1),
-    limit: int = Query(20, ge=1, le=100),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get list of all public bots with pagination
-    """
-    query = db.query(Kind).filter(Kind.user_id == 0, Kind.kind == "Bot")
-    total = query.count()
-
-    # Apply SQL-level pagination
-    paginated_bots = (
-        query.order_by(Kind.updated_at.desc())
-        .offset((page - 1) * limit)
-        .limit(limit)
-        .all()
-    )
-
-    items = []
-    for bot in paginated_bots:
-        ghost_name, shell_name, model_name = _get_bot_ref_info(bot)
-        items.append(
-            PublicBotResponse(
-                id=bot.id,
-                name=bot.name,
-                namespace=bot.namespace,
-                display_name=_get_bot_display_name(bot),
-                bot_json=bot.json,
-                is_active=bot.is_active,
-                created_at=bot.created_at,
-                updated_at=bot.updated_at,
-                ghost_name=ghost_name,
-                shell_name=shell_name,
-                model_name=model_name,
-            )
-        )
-
-    return PublicBotListResponse(total=total, items=items)
-
-
-@router.post(
-    "/public-bots",
-    response_model=PublicBotResponse,
-    status_code=status.HTTP_201_CREATED,
-)
-async def create_public_bot(
-    bot_data: PublicBotCreate,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Create a new public bot (admin only).
-    All resources referenced by the bot (Ghost, Shell, Model) must be public (user_id=0).
-    """
-    # Check if bot with same name and namespace already exists
-    existing_bot = (
-        db.query(Kind)
-        .filter(
-            Kind.user_id == 0,
-            Kind.kind == "Bot",
-            Kind.name == bot_data.name,
-            Kind.namespace == bot_data.namespace,
-        )
-        .first()
-    )
-    if existing_bot:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"Public bot '{bot_data.name}' already exists in namespace '{bot_data.namespace}'",
-        )
-
-    # Validate resource references
-    is_valid, error_message = _validate_bot_resource_references(db, bot_data.bot_json)
-    if not is_valid:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=error_message,
-        )
-
-    new_bot = Kind(
-        user_id=0,
-        kind="Bot",
-        name=bot_data.name,
-        namespace=bot_data.namespace,
-        json=bot_data.bot_json,
-        is_active=True,
-    )
-    db.add(new_bot)
-    db.commit()
-    db.refresh(new_bot)
-
-    ghost_name, shell_name, model_name = _get_bot_ref_info(new_bot)
-    return PublicBotResponse(
-        id=new_bot.id,
-        name=new_bot.name,
-        namespace=new_bot.namespace,
-        display_name=_get_bot_display_name(new_bot),
-        bot_json=new_bot.json,
-        is_active=new_bot.is_active,
-        created_at=new_bot.created_at,
-        updated_at=new_bot.updated_at,
-        ghost_name=ghost_name,
-        shell_name=shell_name,
-        model_name=model_name,
-    )
-
-
-@router.put("/public-bots/{bot_id}", response_model=PublicBotResponse)
-async def update_public_bot(
-    bot_data: PublicBotUpdate,
-    bot_id: int = Path(..., description="Bot ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Update a public bot (admin only).
-    All resources referenced by the bot (Ghost, Shell, Model) must be public (user_id=0).
-    """
-    bot = (
-        db.query(Kind)
-        .filter(Kind.id == bot_id, Kind.user_id == 0, Kind.kind == "Bot")
-        .first()
-    )
-    if not bot:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Public bot with id {bot_id} not found",
-        )
-
-    # Check name uniqueness if being changed
-    if bot_data.name and bot_data.name != bot.name:
-        namespace = bot_data.namespace or bot.namespace
-        existing_bot = (
-            db.query(Kind)
-            .filter(
-                Kind.user_id == 0,
-                Kind.kind == "Bot",
-                Kind.name == bot_data.name,
-                Kind.namespace == namespace,
-            )
-            .first()
-        )
-        if existing_bot:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail=f"Public bot '{bot_data.name}' already exists in namespace '{namespace}'",
-            )
-
-    # Validate resource references if json is being updated
-    if bot_data.bot_json:
-        is_valid, error_message = _validate_bot_resource_references(
-            db, bot_data.bot_json
-        )
-        if not is_valid:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail=error_message,
-            )
-
-    # Update fields
-    if bot_data.name is not None:
-        bot.name = bot_data.name
-    if bot_data.namespace is not None:
-        bot.namespace = bot_data.namespace
-    if bot_data.bot_json is not None:
-        bot.json = bot_data.bot_json
-    if bot_data.is_active is not None:
-        bot.is_active = bot_data.is_active
-
-    db.commit()
-    db.refresh(bot)
-
-    ghost_name, shell_name, model_name = _get_bot_ref_info(bot)
-    return PublicBotResponse(
-        id=bot.id,
-        name=bot.name,
-        namespace=bot.namespace,
-        display_name=_get_bot_display_name(bot),
-        bot_json=bot.json,
-        is_active=bot.is_active,
-        created_at=bot.created_at,
-        updated_at=bot.updated_at,
-        ghost_name=ghost_name,
-        shell_name=shell_name,
-        model_name=model_name,
-    )
-
-
-@router.delete("/public-bots/{bot_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_public_bot(
-    bot_id: int = Path(..., description="Bot ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Delete a public bot (admin only).
-    Deletion is blocked if any public teams (active or inactive) reference this bot.
-    """
-    bot = (
-        db.query(Kind)
-        .filter(Kind.id == bot_id, Kind.user_id == 0, Kind.kind == "Bot")
-        .first()
-    )
-    if not bot:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Public bot with id {bot_id} not found",
-        )
-
-    # Check if any public teams (active or inactive) reference this bot
-    teams_using_bot = (
-        db.query(Kind)
-        .filter(
-            Kind.user_id == 0,
-            Kind.kind == "Team",
-        )
-        .all()
-    )
-
-    referencing_teams = []
-    for team in teams_using_bot:
-        spec = team.json.get("spec", {}) if team.json else {}
-        members = spec.get("members", [])
-        for member in members:
-            bot_ref = member.get("botRef", {})
-            if (
-                bot_ref.get("name") == bot.name
-                and bot_ref.get("namespace", "default") == bot.namespace
-            ):
-                referencing_teams.append(team.name)
-                break
-
-    if referencing_teams:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"Cannot delete bot '{bot.name}' because it is referenced by public teams: {', '.join(referencing_teams)}",
-        )
-
-    db.delete(bot)
-    db.commit()
-
-    return None
-
-
-# ==================== Public Ghost Management Endpoints ====================
-
-
-def _get_ghost_display_name(ghost: Kind) -> Optional[str]:
-    """Extract displayName from ghost json metadata"""
-    if ghost.json and isinstance(ghost.json, dict):
-        metadata = ghost.json.get("metadata", {})
-        if isinstance(metadata, dict):
-            display_name = metadata.get("displayName")
-            if display_name and display_name != ghost.name:
-                return display_name
-    return None
-
-
-def _get_ghost_description(ghost: Kind) -> Optional[str]:
-    """Extract description from ghost json spec"""
-    if ghost.json and isinstance(ghost.json, dict):
-        spec = ghost.json.get("spec", {})
-        if isinstance(spec, dict):
-            return spec.get("description")
-    return None
-
-
-@router.get("/public-ghosts", response_model=PublicGhostListResponse)
-async def list_public_ghosts(
-    page: int = Query(1, ge=1),
-    limit: int = Query(20, ge=1, le=100),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get list of all public ghosts with pagination
-    """
-    query = db.query(Kind).filter(Kind.user_id == 0, Kind.kind == "Ghost")
-    total = query.count()
-
-    # Apply SQL-level pagination
-    paginated_ghosts = (
-        query.order_by(Kind.updated_at.desc())
-        .offset((page - 1) * limit)
-        .limit(limit)
-        .all()
-    )
-
-    return PublicGhostListResponse(
-        total=total,
-        items=[
-            PublicGhostResponse(
-                id=ghost.id,
-                name=ghost.name,
-                namespace=ghost.namespace,
-                display_name=_get_ghost_display_name(ghost),
-                description=_get_ghost_description(ghost),
-                ghost_json=ghost.json,
-                is_active=ghost.is_active,
-                created_at=ghost.created_at,
-                updated_at=ghost.updated_at,
-            )
-            for ghost in paginated_ghosts
-        ],
-    )
-
-
-@router.post(
-    "/public-ghosts",
-    response_model=PublicGhostResponse,
-    status_code=status.HTTP_201_CREATED,
-)
-async def create_public_ghost(
-    ghost_data: PublicGhostCreate,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Create a new public ghost (admin only).
-    """
-    # Check if ghost with same name and namespace already exists
-    existing_ghost = (
-        db.query(Kind)
-        .filter(
-            Kind.user_id == 0,
-            Kind.kind == "Ghost",
-            Kind.name == ghost_data.name,
-            Kind.namespace == ghost_data.namespace,
-        )
-        .first()
-    )
-    if existing_ghost:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"Public ghost '{ghost_data.name}' already exists in namespace '{ghost_data.namespace}'",
-        )
-
-    new_ghost = Kind(
-        user_id=0,
-        kind="Ghost",
-        name=ghost_data.name,
-        namespace=ghost_data.namespace,
-        json=ghost_data.ghost_json,
-        is_active=True,
-    )
-    db.add(new_ghost)
-    db.commit()
-    db.refresh(new_ghost)
-
-    return PublicGhostResponse(
-        id=new_ghost.id,
-        name=new_ghost.name,
-        namespace=new_ghost.namespace,
-        display_name=_get_ghost_display_name(new_ghost),
-        description=_get_ghost_description(new_ghost),
-        ghost_json=new_ghost.json,
-        is_active=new_ghost.is_active,
-        created_at=new_ghost.created_at,
-        updated_at=new_ghost.updated_at,
-    )
-
-
-@router.put("/public-ghosts/{ghost_id}", response_model=PublicGhostResponse)
-async def update_public_ghost(
-    ghost_data: PublicGhostUpdate,
-    ghost_id: int = Path(..., description="Ghost ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Update a public ghost (admin only).
-    """
-    ghost = (
-        db.query(Kind)
-        .filter(Kind.id == ghost_id, Kind.user_id == 0, Kind.kind == "Ghost")
-        .first()
-    )
-    if not ghost:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Public ghost with id {ghost_id} not found",
-        )
-
-    # Check name uniqueness if being changed
-    if ghost_data.name and ghost_data.name != ghost.name:
-        namespace = ghost_data.namespace or ghost.namespace
-        existing_ghost = (
-            db.query(Kind)
-            .filter(
-                Kind.user_id == 0,
-                Kind.kind == "Ghost",
-                Kind.name == ghost_data.name,
-                Kind.namespace == namespace,
-            )
-            .first()
-        )
-        if existing_ghost:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail=f"Public ghost '{ghost_data.name}' already exists in namespace '{namespace}'",
-            )
-
-    # Update fields
-    if ghost_data.name is not None:
-        ghost.name = ghost_data.name
-    if ghost_data.namespace is not None:
-        ghost.namespace = ghost_data.namespace
-    if ghost_data.ghost_json is not None:
-        ghost.json = ghost_data.ghost_json
-    if ghost_data.is_active is not None:
-        ghost.is_active = ghost_data.is_active
-
-    db.commit()
-    db.refresh(ghost)
-
-    return PublicGhostResponse(
-        id=ghost.id,
-        name=ghost.name,
-        namespace=ghost.namespace,
-        display_name=_get_ghost_display_name(ghost),
-        description=_get_ghost_description(ghost),
-        ghost_json=ghost.json,
-        is_active=ghost.is_active,
-        created_at=ghost.created_at,
-        updated_at=ghost.updated_at,
-    )
-
-
-@router.delete("/public-ghosts/{ghost_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_public_ghost(
-    ghost_id: int = Path(..., description="Ghost ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Delete a public ghost (admin only).
-    Deletion is blocked if any public bots (active or inactive) reference this ghost.
-    """
-    ghost = (
-        db.query(Kind)
-        .filter(Kind.id == ghost_id, Kind.user_id == 0, Kind.kind == "Ghost")
-        .first()
-    )
-    if not ghost:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Public ghost with id {ghost_id} not found",
-        )
-
-    # Check if any public bots (active or inactive) reference this ghost
-    bots_using_ghost = (
-        db.query(Kind)
-        .filter(
-            Kind.user_id == 0,
-            Kind.kind == "Bot",
-        )
-        .all()
-    )
-
-    referencing_bots = []
-    for bot in bots_using_ghost:
-        spec = bot.json.get("spec", {}) if bot.json else {}
-        ghost_ref = spec.get("ghostRef", {})
-        if (
-            ghost_ref.get("name") == ghost.name
-            and ghost_ref.get("namespace", "default") == ghost.namespace
-        ):
-            referencing_bots.append(bot.name)
-
-    if referencing_bots:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"Cannot delete ghost '{ghost.name}' because it is referenced by public bots: {', '.join(referencing_bots)}",
-        )
-
-    db.delete(ghost)
-    db.commit()
-
-    return None
-
-
-# ==================== Public Shell Management Endpoints ====================
-
-
-def _get_shell_display_name(shell: Kind) -> Optional[str]:
-    """Extract displayName from shell json metadata"""
-    if shell.json and isinstance(shell.json, dict):
-        metadata = shell.json.get("metadata", {})
-        if isinstance(metadata, dict):
-            display_name = metadata.get("displayName")
-            if display_name and display_name != shell.name:
-                return display_name
-    return None
-
-
-def _get_shell_type(shell: Kind) -> Optional[str]:
-    """Extract shellType from shell json spec"""
-    if shell.json and isinstance(shell.json, dict):
-        spec = shell.json.get("spec", {})
-        if isinstance(spec, dict):
-            return spec.get("shellType")
-    return None
-
-
-@router.get("/public-shells", response_model=PublicShellListResponse)
-async def list_public_shells(
-    page: int = Query(1, ge=1),
-    limit: int = Query(20, ge=1, le=100),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Get list of all public shells with pagination
-    """
-    query = db.query(Kind).filter(Kind.user_id == 0, Kind.kind == "Shell")
-    total = query.count()
-
-    # Apply SQL-level pagination
-    paginated_shells = (
-        query.order_by(Kind.updated_at.desc())
-        .offset((page - 1) * limit)
-        .limit(limit)
-        .all()
-    )
-
-    return PublicShellListResponse(
-        total=total,
-        items=[
-            PublicShellResponse(
-                id=shell.id,
-                name=shell.name,
-                namespace=shell.namespace,
-                display_name=_get_shell_display_name(shell),
-                shell_type=_get_shell_type(shell),
-                shell_json=shell.json,
-                is_active=shell.is_active,
-                created_at=shell.created_at,
-                updated_at=shell.updated_at,
-            )
-            for shell in paginated_shells
-        ],
-    )
-
-
-@router.post(
-    "/public-shells",
-    response_model=PublicShellResponse,
-    status_code=status.HTTP_201_CREATED,
-)
-async def create_public_shell(
-    shell_data: PublicShellCreate,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Create a new public shell (admin only).
-    """
-    # Check if shell with same name and namespace already exists
-    existing_shell = (
-        db.query(Kind)
-        .filter(
-            Kind.user_id == 0,
-            Kind.kind == "Shell",
-            Kind.name == shell_data.name,
-            Kind.namespace == shell_data.namespace,
-        )
-        .first()
-    )
-    if existing_shell:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"Public shell '{shell_data.name}' already exists in namespace '{shell_data.namespace}'",
-        )
-
-    new_shell = Kind(
-        user_id=0,
-        kind="Shell",
-        name=shell_data.name,
-        namespace=shell_data.namespace,
-        json=shell_data.shell_json,
-        is_active=True,
-    )
-    db.add(new_shell)
-    db.commit()
-    db.refresh(new_shell)
-
-    return PublicShellResponse(
-        id=new_shell.id,
-        name=new_shell.name,
-        namespace=new_shell.namespace,
-        display_name=_get_shell_display_name(new_shell),
-        shell_type=_get_shell_type(new_shell),
-        shell_json=new_shell.json,
-        is_active=new_shell.is_active,
-        created_at=new_shell.created_at,
-        updated_at=new_shell.updated_at,
-    )
-
-
-@router.put("/public-shells/{shell_id}", response_model=PublicShellResponse)
-async def update_public_shell(
-    shell_data: PublicShellUpdate,
-    shell_id: int = Path(..., description="Shell ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Update a public shell (admin only).
-    """
-    shell = (
-        db.query(Kind)
-        .filter(Kind.id == shell_id, Kind.user_id == 0, Kind.kind == "Shell")
-        .first()
-    )
-    if not shell:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Public shell with id {shell_id} not found",
-        )
-
-    # Check name uniqueness if being changed
-    if shell_data.name and shell_data.name != shell.name:
-        namespace = shell_data.namespace or shell.namespace
-        existing_shell = (
-            db.query(Kind)
-            .filter(
-                Kind.user_id == 0,
-                Kind.kind == "Shell",
-                Kind.name == shell_data.name,
-                Kind.namespace == namespace,
-            )
-            .first()
-        )
-        if existing_shell:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail=f"Public shell '{shell_data.name}' already exists in namespace '{namespace}'",
-            )
-
-    # Update fields
-    if shell_data.name is not None:
-        shell.name = shell_data.name
-    if shell_data.namespace is not None:
-        shell.namespace = shell_data.namespace
-    if shell_data.shell_json is not None:
-        shell.json = shell_data.shell_json
-    if shell_data.is_active is not None:
-        shell.is_active = shell_data.is_active
-
-    db.commit()
-    db.refresh(shell)
-
-    return PublicShellResponse(
-        id=shell.id,
-        name=shell.name,
-        namespace=shell.namespace,
-        display_name=_get_shell_display_name(shell),
-        shell_type=_get_shell_type(shell),
-        shell_json=shell.json,
-        is_active=shell.is_active,
-        created_at=shell.created_at,
-        updated_at=shell.updated_at,
-    )
-
-
-@router.delete("/public-shells/{shell_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_public_shell(
-    shell_id: int = Path(..., description="Shell ID"),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_admin_user),
-):
-    """
-    Delete a public shell (admin only).
-    Deletion is blocked if any public bots (active or inactive) reference this shell.
-    """
-    shell = (
-        db.query(Kind)
-        .filter(Kind.id == shell_id, Kind.user_id == 0, Kind.kind == "Shell")
-        .first()
-    )
-    if not shell:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Public shell with id {shell_id} not found",
-        )
-
-    # Check if any public bots (active or inactive) reference this shell
-    bots_using_shell = (
-        db.query(Kind)
-        .filter(
-            Kind.user_id == 0,
-            Kind.kind == "Bot",
-        )
-        .all()
-    )
-
-    referencing_bots = []
-    for bot in bots_using_shell:
-        spec = bot.json.get("spec", {}) if bot.json else {}
-        shell_ref = spec.get("shellRef", {})
-        if (
-            shell_ref.get("name") == shell.name
-            and shell_ref.get("namespace", "default") == shell.namespace
-        ):
-            referencing_bots.append(bot.name)
-
-    if referencing_bots:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"Cannot delete shell '{shell.name}' because it is referenced by public bots: {', '.join(referencing_bots)}",
-        )
-
-    db.delete(shell)
-    db.commit()
-
-    return None
